requirements.txt
einops==0.8.0
tensorflow==2.16.1
pyyaml==6.0.1
tqdm==4.67.1